bug fixed - şifre değiştirme

+++ Core/FitTakip.Application/DTOs/AdminDto.cs
@@ -4,7 +4,7 @@ namespace FitTakip.Application.DTOs;
 
 public class AdminDto
 {
-    public int AdminId { get; set; }
+    public long AdminId { get; set; }
     public string Ad { get; set; } = null!;
     public string Soyad { get; set; } = null!;
     public string Rol { get; set; } = null!;
