+++ Core/FitTakip.Application/DTOs/GelirDto.cs
@@ -4,8 +4,8 @@ namespace FitTakip.Application.DTOs;
 
 public class GelirDto
 {
-    public int GelirId { get; set; }
-    public int IsletmeId { get; set; }
+    public long GelirId { get; set; }
+    public long IsletmeId { get; set; }
     public string Aciklama { get; set; } = null!;
     public DateTime Tarih { get; set; }
     public decimal Tutar { get; set; }
