+++ Core/FitTakip.Application/Interfaces/Repositories/IGelirRepository.cs
@@ -5,7 +5,7 @@ namespace FitTakip.Application.Interfaces.Services;
 
 public interface IGelirRepository
 {
-    Task<decimal> TumGelirlerToplamiAsync(int IsletmeId);
-    Task<List<Gelir>> GelirleriTariheGöreGetirPaginationAsync(int IsletmeId, DateTime BaslangicTarihi, DateTime BitisTarihi, int Baslangic, int Adet);
-    Task<decimal> GelirleriTariheGöreToplaAsync(int IsletmeId, DateTime BaslangicTarihi, DateTime BitisTarihi);
+    Task<decimal> TumGelirlerToplamiAsync(long IsletmeId);
+    Task<List<Gelir>> GelirleriTariheGöreGetirPaginationAsync(long IsletmeId, DateTime BaslangicTarihi, DateTime BitisTarihi, int Baslangic, int Adet);
+    Task<decimal> GelirleriTariheGöreToplaAsync(long IsletmeId, DateTime BaslangicTarihi, DateTime BitisTarihi);
 }
