+++ Core/FitTakip.Application/Interfaces/Repositories/IOlcumRepository.cs
@@ -5,5 +5,5 @@ namespace FitTakip.Application.Interfaces.Repositories;
 
 public interface IOlcumRepository
 {
-    Task<List<Olcum?>> UyeyeAitOlcumleriGetirAsync(int UyeId, int Baslangic, int Adet);
+    Task<List<Olcum?>> UyeyeAitOlcumleriGetirAsync(long UyeId, int Baslangic, int Adet);
 }
