+++ Core/FitTakip.Application/Interfaces/Repositories/IPaketRepository.cs
@@ -5,5 +5,5 @@ namespace FitTakip.Application.Interfaces.Repositories;
 
 public interface IPaketRepository
 {
-    Task<List<Paket>> IsletmeyeAitPaketleriGetirAsync(int IsletmeId);
+    Task<List<Paket>> IsletmeyeAitPaketleriGetirAsync(long IsletmeId);
 }
