+++ Core/FitTakip.Application/Interfaces/Repositories/IRepository.cs
@@ -6,6 +6,6 @@ public interface IRepository<T>
 {
     Task CreateAsync(T entity);
     Task<List<T>> GetAllAsync();
-    Task<T?> GetByIdAsync(int Id);
+    Task<T?> GetByIdAsync(long Id);
     Task UpdateAsync(T entity);
 }
