+++ Core/FitTakip.Application/Parametre/AdminGuncelleParametre.cs
@@ -4,7 +4,7 @@ namespace FitTakip.Application.Parametre;
 
 public class AdminGuncelleParametre
 {
-    public int AdminId { get; set; }
+    public long AdminId { get; set; }
     public string? Ad { get; set; } = null!;
     public string? Soyad { get; set; } = null!;
     public string? TelefonNo { get; set; }
