+++ Core/FitTakip.Application/Parametre/EgitmenGuncelleParametre.cs
@@ -4,7 +4,7 @@ namespace FitTakip.Application.Parametre;
 
 public class EgitmenGuncelleParametre
 {
-    public int EgitmenId { get; set; }
+    public long EgitmenId { get; set; }
     public string? Ad { get; set; }  // Isletme, Egitmen ve Uye ortak
     public string? Soyad { get; set; }  // Egitmen ve Uye için geçerli
     public string? TelefonNo { get; set; } // Ortak
