+++ Core/FitTakip.Application/Parametre/EgitmenOlusturParametre.cs
@@ -9,5 +9,5 @@ public class EgitmenOlusturParametre
     public string? TelefonNo { get; set; } // Ortak
     public string KullaniciAdi { get; set; } = null!; // Ortak
     public string Sifre { get; set; } = null!; // Ortak
-    public int IsletmeId { get; set; } // Egitmen ve Uye için geçerli
+    public long IsletmeId { get; set; } // Egitmen ve Uye için geçerli
 }
