+++ Core/FitTakip.Application/Parametre/GelirOlusturParametre.cs
@@ -4,7 +4,7 @@ namespace FitTakip.Application.Parametre;
 
 public class GelirOlusturParametre
 {
-    public int IsletmeId { get; set; }
+    public long IsletmeId { get; set; }
     public string Aciklama { get; set; } = null!;
     public DateTime Tarih { get; set; }
     public decimal Tutar { get; set; }
