+++ Core/FitTakip.Application/Parametre/IsletmeGuncelleParametre.cs
@@ -4,7 +4,7 @@ namespace FitTakip.Application.Parametre;
 
 public class IsletmeGuncelleParametre
 {
-    public int IsletmeId { get; set; }
+    public long IsletmeId { get; set; }
     public string? Ad { get; set; }
     public string? TelefonNo { get; set; }
     public string? KullaniciAdi { get; set; }
