+++ Core/FitTakip.Application/Parametre/IsletmeyeKullanımSuresiEkleParametre.cs
@@ -4,6 +4,6 @@ namespace FitTakip.Application.Parametre;
 
 public class IsletmeyeKullanımSuresiEkleParametre
 {
-    public int IsletmeId { get; set; }
+    public long IsletmeId { get; set; }
     public int EklenecekYil { get; set; }
 }
