+++ Core/FitTakip.Application/Parametre/PaketEkleParametre.cs
@@ -4,7 +4,7 @@ namespace FitTakip.Application.Parametre;
 
 public class PaketEkleParametre
 {
-    public int IsletmeId { get; set; }
+    public long IsletmeId { get; set; }
     public string Aciklama { get; set; } = null!;
     public decimal Tutar { get; set; }
     public int DersSayisi { get; set; }
