+++ Core/FitTakip.Application/Parametre/RandevuOlusturParametre.cs
@@ -4,8 +4,8 @@ namespace FitTakip.Application.Parametre;
 
 public class RandevuOlusturParametre
 {
-    public int UyeId { get; set; }
-    public int EgitmenId { get; set; }
+    public long UyeId { get; set; }
+    public long EgitmenId { get; set; }
     public DateTime Tarih { get; set; }
     public string? Aciklama { get; set; }
 }
