+++ Core/FitTakip.Application/Parametre/UyeGuncelleParametre.cs
@@ -4,7 +4,7 @@ namespace FitTakip.Application.Parametre;
 
 public class UyeGuncelleParametre
 {
-    public int UyeId { get; set; }
+    public long UyeId { get; set; }
     public string? Ad { get; set; } = null!;
     public string? Soyad { get; set; } = null!;
     public string? TelefonNo { get; set; }
