+++ Core/FitTakip.Application/Parametre/UyeOlusturParametre.cs
@@ -7,7 +7,7 @@ public class UyeOlusturParametre
     public string Ad { get; set; } = null!;
     public string Soyad { get; set; } = null!; 
     public string? TelefonNo { get; set; }
-    public int PaketId { get; set; } 
-    public int IsletmeId { get; set; } 
-    public int EgitmenId { get; set; } 
+    public long PaketId { get; set; } 
+    public long IsletmeId { get; set; } 
+    public long EgitmenId { get; set; } 
 }
