+++ Core/FitTakip.Domain/Entities/Egitmen.cs
@@ -4,13 +4,13 @@ namespace FitTakip.Domain.Entities;
 
 public class Egitmen
 {
-    public int EgitmenId { get; set; }
+    public long EgitmenId { get; set; }
     public string Ad { get; set; } = null!;
     public string Soyad { get; set; } = null!;
     public string? TelefonNo { get; set; }
     public string KullaniciAdi { get; set; } = null!;
     public string SifreKarmasi { get; set; } = null!;
-    public int IsletmeId { get; set; }
+    public long IsletmeId { get; set; }
     public bool AktifMi { get; set; }
 
     public Isletme Isletme { get; set; } = null!;
