+++ Core/FitTakip.Domain/Entities/Gider.cs
@@ -4,8 +4,8 @@ namespace FitTakip.Domain.Entities;
 
 public class Gider
 {
-    public int GiderId { get; set; }
-    public int IsletmeId { get; set; }
+    public long GiderId { get; set; }
+    public long IsletmeId { get; set; }
     public string Aciklama { get; set; } = null!;
     public DateTime Tarih { get; set; }
     public decimal Tutar { get; set; }
