+++ Core/FitTakip.Domain/Entities/Olcum.cs
@@ -5,7 +5,7 @@ namespace FitTakip.Domain.Entities;
 public class Olcum
 {
     public long OlcumId { get; set; }
-    public int UyeId { get; set; }
+    public long UyeId { get; set; }
     public DateTime Tarih { get; set; }
     public int Boy { get; set; }
     public decimal Kilo { get; set; }
