+++ Core/FitTakip.Domain/Entities/Randevu.cs
@@ -5,8 +5,8 @@ namespace FitTakip.Domain.Entities;
 public class Randevu
 {
     public long RandevuId { get; set; }
-    public int UyeId { get; set; }
-    public int EgitmenId { get; set; }
+    public long UyeId { get; set; }
+    public long EgitmenId { get; set; }
     public DateTime Tarih { get; set; }
     public string? Aciklama { get; set; }
