+++ Infrastructure/FitTakip.Persistence/Repository/GelirRepository.cs
@@ -15,19 +15,19 @@ public class GelirRepository : IGelirRepository
         _context = context;
     }
 
-    public async Task<decimal> TumGelirlerToplamiAsync(int IsletmeId)
+    public async Task<decimal> TumGelirlerToplamiAsync(long IsletmeId)
     {
         return await _context.Gelirler.Where(w => w.IsletmeId == IsletmeId).SumAsync(s => s.Tutar);
     }
 
-    public async Task<List<Gelir>> GelirleriTariheGöreGetirPaginationAsync(int IsletmeId, DateTime BaslangicTarihi, DateTime BitisTarihi, int Baslangic, int Adet)
+    public async Task<List<Gelir>> GelirleriTariheGöreGetirPaginationAsync(long IsletmeId, DateTime BaslangicTarihi, DateTime BitisTarihi, int Baslangic, int Adet)
     {
         var baslangicTarihi = BaslangicTarihi.Date;
         var bitisTarihi = BitisTarihi.Date.AddHours(23).AddMinutes(59).AddSeconds(59);
         return await _context.Gelirler.Where(w => w.IsletmeId == IsletmeId && w.Tarih >= baslangicTarihi && w.Tarih <= bitisTarihi).OrderByDescending(o => o.Tarih).Skip(Baslangic).Take(Adet).ToListAsync();
     }
 
-    public async Task<decimal> GelirleriTariheGöreToplaAsync(int IsletmeId, DateTime BaslangicTarihi, DateTime BitisTarihi)
+    public async Task<decimal> GelirleriTariheGöreToplaAsync(long IsletmeId, DateTime BaslangicTarihi, DateTime BitisTarihi)
     {
         var baslangicTarihi = BaslangicTarihi.Date;
         var bitisTarihi = BitisTarihi.Date.AddHours(23).AddMinutes(59).AddSeconds(59);
