+++ Infrastructure/FitTakip.Persistence/Repository/OlcumRepository.cs
@@ -15,7 +15,7 @@ public class OlcumRepository : IOlcumRepository
         _context = context;
     }
 
-    public async Task<List<Olcum>> UyeyeAitOlcumleriGetirAsync(int UyeId, int Baslangic, int Adet)
+    public async Task<List<Olcum>> UyeyeAitOlcumleriGetirAsync(long UyeId, int Baslangic, int Adet)
     {
         return await _context.Olcumler.AsNoTracking().Where(w => w.UyeId == UyeId).Include(i => i.Uye).OrderByDescending(o => o.Tarih).Skip(Baslangic).Take(Adet).ToListAsync();
     }
